Add login.html for user authentication
This commit adds a login page (login.html) to the website, which includes a form for users to enter their username and password. Upon submission, the form simulates user authentication. If the provided credentials match the hardcoded values, the user is redirected to the main page; otherwise, an error message is displayed.


Signed-off-by: MARTOOTECHHUB <martootechhub@gmail.com>